refactor the database
Enahance reusability and more
