+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod auth;
 pub mod cli;
 pub mod config;
 pub mod constants;
+pub mod database;
 pub mod schedule;
 pub mod twitter;
 pub mod usage;
